+++ timer/timer.h
@@ -11,10 +11,10 @@ void timer_init();
  * Returns the number of milliseconds which have elapsed since the 
  * last time timer_init() was called.  Overflows after about 49 days.
  */
-unsigned long long timer_millis();
+uint64_t timer_millis();
 
 /*
  * Returns the number of microseconds which have elapsed since the 
  * last time timer_init() was called.  Overflows after about 71 minutes.
  */
-unsigned long long timer_micros();
+uint64_t timer_micros();
